[CHORE] #125 - 장바구니 열기 중복되는 코드 제거
